+++ configurator-service/src/test/java/com/frejdh/util/environment/test/helper/ToListTests.java
@@ -1,7 +1,5 @@
 package com.frejdh.util.environment.test.helper;
 
-import static org.junit.Assert.assertEquals;
-
 public class ToListTests extends AbstractTests {
 
 //	@Test